start wiring up actions, game loop, and game init
Update progressGame function logic to work with the new turnCount
parameter.

Adds a BeatManager class responsible for managing the game clock and
music. Wires this up to the webocket client with a callback to call
tick on every beat.

Game state now initializes from server event with the shared player
id. Add a comment with thougts on how this can be expanded for
multiplayer awareness.

Expose an "act" function from the client for the App component to use
to send player actions.

Nit -- it kinda bothers me that WebSocketClient/client is really just
the glue of everything to make the game work.. feels like it should be
named differently and live somewhere else...

Removes seemingly accidentally committed node gunk within the go
server directory.